# src/main/java/td/agricoles/agricol/dto/request/CollectiveIdentifiersAssignment.java
package td.agricoles.agricol.dto.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CollectiveIdentifiersAssignment {
    private String number;
    private String name;

}