# src/main/java/td/agricoles/agricol/dto/response/Collectivity.java
package td.agricoles.agricol.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Collectivity {
    private String id;
    private String number;
    private String name;
    private String location;
    private CollectivityStructure structure;
    private List<Member> members;
}